Fixed fragment device bugs
