+++ app/src/main/java/com/apptec/registrateapp/mainactivity/MainPresenterImpl.java
@@ -24,11 +24,5 @@ public class MainPresenterImpl {
         mainInteractor.initializeDeviceVerification(isNeedRegisterDevice);
     }
 
-    public void saveThisDevice(String name, String model, MutableLiveData<Boolean> isNeedRegisterDevice) {
-        /**
-         * Calling the interactor
-         */
-        mainInteractor.saveThisDevice(name, model, isNeedRegisterDevice);
-    }
 
 }
