--- app/src/main/java/com/apptec/registrateapp/mainactivity/fdevice/DevicePresenter.java
@@ -1,15 +1,0 @@
-package com.apptec.registrateapp.mainactivity.fdevice;
-
-import com.apptec.registrateapp.models.Device;
-
-import java.util.ArrayList;
-
-public interface DevicePresenter {
-    /**
-     * Device Presenter
-     * */
-
-    void getDevices();                             // To the interactor
-    void showDevices(ArrayList<Device> devices);   // From the interactor to the view
-
-}
